Add files via upload
This is the first version of the ESPeciallyKerbal branch, which I have decided to use, after problems with using an SD card on the Sprig to host a web UI. 
The ESPeciallyKerbal companion program allows users to use an ESP32 to take advantage of the web APIs exposed by the Sprig, which allows you to use a web UI to control Kerbal Space program. It also includes support for an authentication token to allow use of certain risky actions, like the abort function.